+++ lib/src/helper.dart
@@ -1,3 +1,7 @@
+import 'dart:io';
+import 'package:flutter/services.dart';
+import 'package:http/http.dart' as http;
+
 import 'package:json_annotation/json_annotation.dart';
 part 'helper.g.dart';
 
@@ -63,7 +67,7 @@ class EpubDisplaySettings {
   int fontSize;
   EpubSpread spread;
   EpubFlow flow;
-  EpubDefaultDirection? defaultDirection;
+  EpubDefaultDirection defaultDirection;
   bool allowScriptedContent;
   EpubManager manager;
 
@@ -75,7 +79,7 @@ class EpubDisplaySettings {
     this.spread = EpubSpread.auto,
     this.flow = EpubFlow.scrolled,
     this.allowScriptedContent = false,
-    this.defaultDirection,
+    this.defaultDirection = EpubDefaultDirection.ltr,
     this.snap = false,
     this.manager = EpubManager.continuous,
   });
@@ -110,12 +114,75 @@ enum EpubManager {
   // epub
 }
 
+///Epub text selection callback object
 class EpubTextSelection {
+  /// The selected text
   final String selectedText;
+
+  /// The cfi string of the selected text
   final String selectionCfi;
 
   EpubTextSelection({
     required this.selectedText,
     required this.selectionCfi,
   });
+}
+
+/// Epub file source
+class EpubSource {
+  // final Uint8List epubData;
+  final Future<Uint8List> epubData;
+
+  EpubSource._({required this.epubData});
+
+  ///Loading from a file
+  factory EpubSource.fromFile(File file) {
+    return EpubSource._(epubData: file.readAsBytes());
+  }
+
+  ///load from a url with optional headers
+  factory EpubSource.fromUrl(String url, {Map<String, String>? headers}) {
+    return EpubSource._(epubData: _downloadFile(url, headers: headers));
+  }
+
+  ///load from assets
+  factory EpubSource.fromAsset(String assetPath) {
+    return EpubSource._(epubData: _getAssetData(assetPath));
+  }
+
+  static Future<Uint8List> _getAssetData(assetPath) {
+    final byteData = rootBundle.load(assetPath);
+    return byteData.then((val) => val.buffer.asUint8List());
+  }
+
+  static Future<Uint8List> _downloadFile(String url,
+      {Map<String, String>? headers}) async {
+    try {
+      final response = await http.get(Uri.parse(url), headers: headers);
+
+      if (response.statusCode == 200) {
+        return response.bodyBytes;
+      } else {
+        throw Exception('Failed to download file from URL');
+      }
+    } catch (e) {
+      throw Exception('Failed to download file from URL, $e');
+    }
+  }
+}
+
+@JsonSerializable(explicitToJson: true)
+
+///Epub text extraction callback object
+class EpubTextExtractRes {
+  String? text;
+  String? cfiRange;
+
+  EpubTextExtractRes({
+    this.text,
+    this.cfiRange,
+  });
+  factory EpubTextExtractRes.fromJson(Map<String, dynamic> json) =>
+      _$EpubTextExtractResFromJson(json);
+  Map<String, dynamic> toJson() => _$EpubTextExtractResToJson(this);
 }